requirements.txt
flask>=2.3.0
playwright>=1.40.0
requests>=2.31.0
colorama>=0.4.6
qiniu>=7.17.0
PySocks>=1.7.1